fix(api-server): handle path names with 'graphql' (#12055)
When this repo changed name from 'redwood' to 'graphql' GitHub also
changed the path for the project on its runners. With the new repo name
the paths look like this:

```js
serverFunctions: [
  '/home/runner/work/graphql/graphql/packages/api-server/src/__tests__/fixtures/redwood-app/api/dist/functions/env.js',
  '/home/runner/work/graphql/graphql/packages/api-server/src/__tests__/fixtures/redwood-app/api/dist/functions/graphql.js',
  '/home/runner/work/graphql/graphql/packages/api-server/src/__tests__/fixtures/redwood-app/api/dist/functions/health.js',
  '/home/runner/work/graphql/graphql/packages/api-server/src/__tests__/fixtures/redwood-app/api/dist/functions/hello.js',
  '/home/runner/work/graphql/graphql/packages/api-server/src/__tests__/fixtures/redwood-app/api/dist/functions/noHandler.js',
  '/home/runner/work/graphql/graphql/packages/api-server/src/__tests__/fixtures/redwood-app/api/dist/functions/nested/nested.js'
]
```

Previously the code was looking for paths that included `'graphql'` to
find the graphql function. The problem is that the paths now all include
`'graphql'`. So this PR changes the code to look for paths that ends
with `'graphql.js'` to find the graphql function